[Change] CSSの切り替えが即時反映されるよう変更
