+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "laf",
-    "version": "5.4.7",
+    "version": "5.5.1-beta.1",
     "description": "Lite and Fast Krunker Client",
     "main": "app/main.js",
     "scripts": {
